requirements.txt
python-telegram-bot[job-queue]==20.7
google-generativeai==0.8.3
firebase-admin==6.6.0
gTTS==2.5.4
python-dotenv==1.0.1
APScheduler==3.10.4
